+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * Add `background-attachment` property to SMACSS property sort order list
 * Add support for CSS Logical Properties to `PropertySpelling`
 * Add `initial` to list of special values in `VariableForProperty`
+* Add `::-webkit-search-decoration`/`::-webkit-search-results-decoration`
+  pseudo elements to `PseudoElement`
 
 ## 0.43.2